Fix AttributeError when creating new Icarus worker.
Prior to this commit, I was unable to add an Icarus worker via the
web interface.  I got an error dialog which said:

  AttributeError: 'IcarusWorker' object has no attribute 'baudrate'

This appears to be a regression introduced in commit 4a8cbaf4.  The
same bug is also present in SimpleRS23Worker, so I fixed it there
as well.